refactor(overloaded_visitor_t): renames to functor_visitor_t, creates overloaded_visitor_t
