+++ include/soccer-common/Utils/Utils.h
@@ -11,7 +11,7 @@
 #include "Singleton/Singleton.h"
 #include "SharedValue/SharedValue.h"
 #include "SharedWrapper/SharedWrapper.h"
-#include "overloaded_visitor_t/overloaded_visitor_t.h"
+#include "functor_visitor_t/functor_visitor_t.h"
 #include "IthSender/IthSender.h"
 #include "DeleteLaterDeleter/DeleteLaterDeleter.h"
 #include "Protobuf/Protobuf.h"
@@ -19,5 +19,6 @@
 #include "is_in/is_in.h"
 #include "OStream/OStream.h"
 #include "utility/utility.h"
+#include "overloaded_visitor_t/overloaded_visitor_t.h"
 
 #endif // SOCCER_COMMON_UTILS_H
