--- include/soccer-common/Utils/overloaded_visitor_t/overloaded_visitor_t.cpp
@@ -1,5 +1,0 @@
-//
-// Created by jose-cruz on 15/02/2021.
-//
-
-#include "overloaded_visitor_t.h"
